Refactor EpisodeDetails: Move PlayerDanmakuState to `.ui.subject.episode.danmaku`
